change the type of 'aloha_bandwidth_' from int to double

+++ satellite/PureAloha.h
@@ -21,6 +21,6 @@ protected:
 	double send_timeout_;	// Set in OTcl-- time out after this interval
 	double end_of_contention_; // Saves time that contention will be over
 	//double bit_duration_;
-	int aloha_bandwidth_;
+	double aloha_bandwidth_;
 	int pkt_bit_length_;
 };
